Add session handling and update dependencies
This commit introduces session handling which includes creating a new Session struct and SessionService in session.go. A new sessions table is also created in sessions.sql to store session data. The "rand" package is introduced to generate random session tokens. Also, gorilla/csrf is now a direct dependency and the users table creation in users.sql is now enclosed with "IF NOT EXISTS" to prevent errors during repeated runs. This update is aimed to bolster security and improve user authentication process.