Enhance user navigation and image handling throughout galleries
This update includes significant enhancements to image handling and user navigation. User redirection post-sign up and sign-in is set to the galleries page for improved navigation. Additionally, code for absolute positioning of images and grid column distribution is added for layout optimization. Image upload and deletion functionality are added to the gallery edit page, and a visual list of current images is displayed. The models now support the checks for content type of files, allowing only png, gif, and jpg files to be uploaded. This improvement ensures better data quality and user experience.

Signed-off-by: Roberto L. Taborda <robert@zelinzky.dev>